Login and log out working

+++ frontend/js/logout.js
@@ -1,12 +1,16 @@
 {
 	console.log("SCRIPT LOG OUT")
 	console.log("Token: " + localStorage.token)
-	const logout = document.getElementById("logout");
 
-	logout.addEventListener("click", (e) => {
-		e.preventDefault();
-		alert("Logout successful!");
-		localStorage.clear();
-		urlRoute({ target: { href: '/' }, preventDefault: () => {} });
-	})
+	if (localStorage.getItem("token"))
+	{
+		const logout = document.getElementById("logout");
+
+		logout.addEventListener("click", (e) => {
+			e.preventDefault();
+			alert("Logout successful!");
+			localStorage.clear();
+			urlRoute({ target: { href: '/' }, preventDefault: () => {} });
+		})
+	}
 }
